pyproject.toml
[project]
name = "mcpagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "arcadepy==1.5.0",
    "Authlib==1.6.0",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "cloudpickle==3.1.1",
    "colorama==0.4.6",
    "cryptography==45.0.4",
    "distro==1.9.0",
    "docstring_parser==0.16",
    "fastapi==0.115.12",
    "google-adk==1.2.1",
    "google-adk-arcade==0.0.2",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.1rc0",
    "google-api-python-client==2.172.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-cloud-aiplatform",
    "google-cloud-appengine-logging==1.6.2",
    "google-cloud-audit-log==0.3.2",
    "google-cloud-bigquery",
    "google-cloud-core==2.4.3",
    "google-cloud-logging==3.12.1",
    "google-cloud-resource-manager==1.14.2",
    "google-cloud-secret-manager==2.24.0",
    "google-cloud-speech==2.33.0",
    "google-cloud-storage==2.19.0",
    "google-cloud-trace==1.16.2",
    "google-crc32c==1.7.1",
    "google-genai==1.20.0",
    "google-generativeai==0.8.5",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "graphviz==0.20.3",
    "greenlet==3.2.3",
    "grpc-google-iam-v1==0.14.2",
    "grpcio==1.73.0",
    "grpcio-status==1.71.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "idna==3.10",
    "importlib_metadata==8.7.0",
    "mcp==1.9.4",
    "numpy==2.3.0",
    "opentelemetry-api==1.34.1",
    "opentelemetry-exporter-gcp-trace==1.9.0",
    "opentelemetry-resourcedetector-gcp==1.9.0a0",
    "opentelemetry-sdk==1.34.1",
    "opentelemetry-semantic-conventions==0.55b1",
    "packaging",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.5",
    "pydantic-settings==2.9.1",
    "pydantic_core==2.33.2",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "PyYAML==6.0.2",
    "requests==2.32.4",
    "rsa==4.9.1",
    "shapely==2.1.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.41",
    "sse-starlette==2.3.6",
    "starlette==0.46.2",
    "streamlit",
    "tenacity==9.1.2",
    "toml==0.10.2",
    "tornado==6.5.1",
    "tqdm==4.67.1",
    "typing-inspection==0.4.1",
    "typing_extensions==4.14.0",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "uritemplate==4.2.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.3",
    "websockets==15.0.1",
    "zipp==3.23.0",
    "langchain-mcp-adapters",
    "fastmcp",
    "reflex>=0.8.0",
    "openai>=1.93.0",
]
